pyproject.toml
[tool.poetry]
name = "nonebot_plugin_rauthman"
version = "2.0.0-alpha.8.post1"
description = "Rule-based authorization manager worked with nonebot2"
authors = ["Lancercmd <lancercmd@gmail.com>"]
license = "GPLv3"
readme = "README.md"
repository = "https://github.com/Lancercmd/nonebot_plugin_rauthman"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-alpha.8"
ujson = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
